requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
reportlab
httpx
python-dotenv==0.19.0
aiofiles
openai-whisper
cffi
soundfile
langchain
pymupdf
huggingface-hub
faiss-cpu
sentence-transformers
cashfree_pg
cashfree_verification